added remaining functions
